b/260769399 Extract email-specific logic from NotificationService (#203)
* Turn MailNotificationService into top-level class
* Move NotificationTemplate to MailNotificationService as it's 
   mostly relevant for formatting email
