+++ sources/src/main/java/com/google/solutions/jitaccess/web/RuntimeEnvironment.java
@@ -319,9 +319,9 @@ public class RuntimeEnvironment {
           this.configuration.smtpPassword.getValue());
       }
 
-      return new NotificationService.MailNotificationService(
+      return new MailNotificationService(
         new SmtpAdapter(secretManagerAdapter, options),
-        new NotificationService.Options(this.configuration.timeZoneForNotifications.getValue()));
+        new MailNotificationService.Options(this.configuration.timeZoneForNotifications.getValue()));
     }
     else {
       return new NotificationService.SilentNotificationService();
